requirements.txt
streamlit
crewai
langchain
langchain-google-genai
langchain-chroma
langchain-text-splitters
langchain-core
chromadb
youtube-transcript-api
python-dotenv
